moved mixedModal to separate component
