+++ src/App.jsx
@@ -5,6 +5,7 @@ import MusicControls from "./components/musicControls/MusicControls";
 import Clock from "./components/clock/Clock";
 import logo from "./assets/logo.png";
 import Mixer from "./components/mixer/Mixer";
+import MixerModal from "./components/mixer/MixerModal";
 import AppContextProvider from "./context/AppContext";
 // import GenreDropdown from "./components/genreDropdown/GenreDropdown";
 import NowPlaying from "./components/nowPlaying/NowPlaying";
@@ -15,7 +16,9 @@ import { RiFullscreenFill, RiFullscreenExitLine, GiTacos } from "./imports/icons
 import Tooltip from "./components/tooltip/Tooltip";
 
 function App() {
-  const [showScene, setShowScene] = useState(false);
+  const [showMixerModal, setShowMixerModal] = useState(false);
+  const [isSleep, setIsSleep] = useState(false);
+
   const handle = useFullScreenHandle();
 
   const handleFullscreen = () => {
@@ -25,6 +28,27 @@ function App() {
       handle.enter();
     }
   };
+
+  useEffect(() => {
+    console.log("From App: button pressed");
+  }, [showMixerModal]);
+
+  let timeout;
+  useEffect(() => {
+    const onMouseMove = () => {
+      setIsSleep(false);
+      clearTimeout(timeout);
+
+      timeout = setTimeout(() => {
+        setIsSleep(true);
+      }, 15000);
+    };
+    document.addEventListener("mousemove", onMouseMove);
+
+    return () => {
+      document.removeEventListener("mousemove", onMouseMove);
+    };
+  }, []);
 
   // Use this when you are ready to redirect away from mobile devices
   // const [screenSize, setScreenSize] = useState(getCurrentDimension());
@@ -54,43 +78,69 @@ function App() {
         <div className="App" id="app">
           <SceneBg />
 
-          <nav id="nav" className="melofi__nav">
-            <div className="melofi__logo">
-              <img src={logo} alt="melofi logo" />
-            </div>
+          {/* Header */}
+          <div
+            className="melofi__header"
+            style={
+              isSleep
+                ? { animation: "slide-up 0.4s forwards" }
+                : { animation: "slide-down 0.4s forwards" }
+            }
+          >
+            <nav id="nav" className="melofi__nav">
+              <div className="melofi__logo">
+                <img
+                  src={logo}
+                  alt="melofi logo"
+                  onClick={() => setShowMixerModal((prev) => !prev)}
+                />
+              </div>
 
-            <div className="melofi__rightSide">
-              {/* GenreDropdown will be a future feature */}
-              {/* <GenreDropdown /> */}
-              <Mixer />
-              <MusicControls />
-              <Scene showScene={showScene} setShowScene={setShowScene} />
-              <Tooltip text={handle.active ? "Exit full screen" : "Enter full screen"}>
-                <div
-                  className="melofi__fullscreen-button"
-                  onClick={handleFullscreen}
-                  style={handle.active ? { outline: "1px solid rgba(254, 165, 57, 0.88)" } : {}}
-                >
-                  {handle.active ? (
-                    <RiFullscreenExitLine size={20} color="white" />
-                  ) : (
-                    <RiFullscreenFill size={20} color="white" />
-                  )}
-                </div>
-              </Tooltip>
-              <Clock />
-            </div>
-          </nav>
+              <div className="melofi__rightSide">
+                {/* GenreDropdown will be a future feature */}
+                {/* <GenreDropdown /> */}
+                <Mixer setShowMixerModal={setShowMixerModal} showMixerModal={showMixerModal} />
+                <MusicControls />
+                <Scene />
+                <Tooltip text={handle.active ? "Exit full screen" : "Enter full screen"}>
+                  <div
+                    className="melofi__fullscreen-button"
+                    onClick={handleFullscreen}
+                    style={handle.active ? { outline: "1px solid rgba(254, 165, 57, 0.88)" } : {}}
+                  >
+                    {handle.active ? (
+                      <RiFullscreenExitLine size={20} color="white" />
+                    ) : (
+                      <RiFullscreenFill size={20} color="white" />
+                    )}
+                  </div>
+                </Tooltip>
+                <Clock />
+              </div>
+            </nav>
+          </div>
 
-          <NowPlaying />
-          <div className="melofi__buyMeATacoLink">
-            <div>
-              <GiTacos size={30} color="var(--color-secondary-white)" />
-              <a href="https://bmc.link/normancade" target="_blank">
-                Buy me a taco
-              </a>
+          {/* Footer */}
+          <div
+            className="melofi__footer"
+            style={
+              isSleep
+                ? { animation: "slide-footer-down 0.4s forwards" }
+                : { animation: "slide-footer-up 0.4s forwards" }
+            }
+          >
+            <NowPlaying />
+            <div className="melofi__buyMeATacoLink">
+              <div>
+                <GiTacos size={30} color="var(--color-secondary-white)" />
+                <a href="https://bmc.link/normancade" target="_blank">
+                  Buy me a taco
+                </a>
+              </div>
             </div>
           </div>
+
+          <MixerModal showModal={showMixerModal} setShowModal={setShowMixerModal} />
         </div>
       </FullScreen>
     </AppContextProvider>
