+++ src/components/mixer/mixerModal.css
@@ -1,0 +1,61 @@
+.melofi__mixerModal {
+    position: absolute;
+    top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
+    margin: auto;
+    background-color: var(--color-primary);
+    width: 378px;
+    height: 530px;
+    padding: 20px 24px;
+    border-radius: 10px;
+    z-index: 100;
+    backdrop-filter: blur(10px);
+    box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
+    overflow: scroll;
+  }
+
+  ::-webkit-scrollbar {
+    width: 0px;
+    background: transparent; /* make scrollbar transparent */
+}
+
+.melofi__mixer-modal-handle { 
+  width: 100%; 
+  height: 20px; 
+  position: absolute; 
+  top: 0; 
+  left: 0; 
+  border-top-left-radius: 10px; 
+  border-top-right-radius: 10px; 
+  cursor: all-scroll;
+}
+
+.melofi__mixer_header {
+    display: flex;
+    justify-content: space-between;
+ }
+
+.melofi__mixer-title {
+    color: var(--color-secondary-white);
+    font-family: var(--font-poppins);
+    font-size: 32px;
+    margin-bottom: 10px;
+}
+
+.melofi__mixer_volume-title {
+    color: var(--color-secondary-white);
+    font-family: var(--font-poppins);
+    font-size: 21px;
+}
+
+.melofi__mixer_reset-button {
+    color: var(--color-secondary);
+    font-family: var(--font-poppins);
+    font-size: 21px;
+    text-align: center;
+    margin-top: 35px;
+    cursor: pointer;
+}
+
