+++ src/index.css
@@ -29,6 +29,9 @@ p {
 a {
   color: inherit; 
   text-decoration: inherit; 
+  -webkit-user-select: none; /* Safari */
+  -ms-user-select: none; /* IE 10 and IE 11 */
+  user-select: none;
 }
 
 img {
